Show errors as notifications, on Logs tab, and mod table (#76)
* Run bat and quit app

* Copying update and running new manager

* Exit updater after running manager

* cd

* woops zip

* Logging progress

* unzip progress

* Refactor file operations into separate service

* add update dir to gitignore

* mod config instead of enabler

* Fix running bat

* Fix bat file path

* Use exit instead of quit

* Show download / unzip progress

* Remove self update stuff temporarily

* Remove self update stuff temporarily

* Circular progress

* Disable install button if loading

* Disable button highlight if loading

* show global progress bar while downloading

* default background for owml row

* glow

* Remove mod.isLoading

* Reset progress to 0

* Cleanup

* Re-add auto update stuff

* Remove leftover isLoading propertie

* Add settings page

* Add setting for close on play

* Use close on play setting

* Option for sending logs to mod manager

* Vertical settings

* Tab icons

* Refactor tabs

* Use list for settings page

* Cleanup

* Refactor settings

* Play icon

* Getting settings from file

* Changing settings file

* One component per input type

* Slider

* Cleanup slider

* Text field

* Remove settings.json

* Write settings file if none exists

* Using mod database URL from settings

* Reset button

* Reset to default adjustments

* Style fixes

* Advanced settings toggle

* Save button on text input

* Add hooks eslint rules

* Rename pages

* Fix eslint

* Implicit coercion rule

* Fix import comment

* Lint rule: useless else

* Remove close on play settin

* run as admin

* Revert "Remove close on play settin"

This reverts commit 8b8afa64bb0cf3955e386e142a212ac761917e3f.

* Run OWML detached

* dried up setting form control

* Disable send logs to manager if manager is set to auto close

* Add update tab

* Fix watcher close crash

* fix tab value

* Update page

* Update page components

* Checking if app is outdated

* Refactor unzipRemoteFile

* Download update on click

* Using node-fetch instead of request for file downloads

* Catch network error on file download

* Cleanup

* Cleanup

* use-update-page hook

* Cleanup

* Notifications provider

* Notifications component

* Show message on update error

* Cleanup

* Notification colors

* Success notification only if success

* Secondary color update icon

* Install update button

* Add app release to app state

* Read app release from database

* Use release url from database

* Prevent running script in places it shouldn't be run

* Cleanup

* Copy install-update.bat to release

* Cleanup

* Update page styling

* Remove redundant manage repo setting

* Fix update path

* Remove axios dependency

* Cleanup

* Update React, use stable version

* Use openPath instead of openItem

* Re-fix react package

* Fix HTML hierarchy in chips

* Cleanup

* Change scroll bar style

* Catch error on getting mod config

* Error when one mod config fails

* Add OWML to local mods

* Prevent unnecessary notification updates

* Prevent repeated messages

* Async saveconfig

* setLocalModMap once

* Back to sync enable disable

* Mark mod as broken

* Show errors in tooltip

* Mod error in notification

* Sending errors to logs

* Manifest attribute errors

* Fix log type

* Fix duplicate log lines

* Errors on remote mod manifest

* Clean

* Refactor

* enable toggle if broken

* Error on database download

* error on mod action

* Delete folder errors

* Open directory errors

* Cleanup

* cleanup

* Fixes

* Cleanup

+++ src/services/files.ts
@@ -20,7 +20,7 @@ export async function downloadFile(
 
   return new Promise((resolve, reject) => {
     if (!response.ok) {
-      reject(response.statusText);
+      reject(`${response.statusText} (${response.status})`);
     }
 
     const totalBytes = parseInt(response.headers.get('content-length') || '0');
@@ -80,7 +80,13 @@ export async function copyFolder(sourcePath: string, targetPath: string) {
 }
 
 export function deleteFolder(folderPath: string) {
-  fs.removeSync(folderPath);
+  if (fs.existsSync(folderPath)) {
+    fs.removeSync(folderPath);
+  } else {
+    throw new Error(
+      `Trying to delete a non-existing directory: "${folderPath}"`,
+    );
+  }
 }
 
 export async function unzipRemoteFile(
